chore (refs T22614): move DpVideoPlayer to client/js.
