+++ demosplan/DemosPlanCoreBundle/Resources/client/js/bundles/publicDetail.js
@@ -14,7 +14,7 @@
 
 import DpPublicDetailNoMap from '@DemosPlanStatementBundle/components/DpPublicDetailNoMap'
 import DpUploadFiles from '@DpJs/components/core/DpUpload/DpUploadFiles'
-import DpVideoPlayer from '@DemosPlanCoreBundle/components/DpVideoPlayer'
+import DpVideoPlayer from '@DpJs/components/core/DpVideoPlayer'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
 import publicStatement from '@DemosPlanStatementBundle/store/PublicStatement'
 import RegisterFlyout from '@DemosPlanCoreBundle/components/RegisterFlyout'
