+++ demosplan/DemosPlanMapBundle/Resources/client/js/bundles/publicDetail.js
@@ -18,7 +18,7 @@ import DpPublicLayerListWrapper from '@DemosPlanMapBundle/components/publicdetai
 import DpPublicSurvey from '@DemosPlanProcedureBundle/components/survey/DpPublicSurvey'
 import DpUnfoldToolbarControl from '@DemosPlanMapBundle/components/publicdetail/controls/DpUnfoldToolbarControl'
 import DpUploadFiles from '@DpJs/components/core/DpUpload/DpUploadFiles'
-import DpVideoPlayer from '@DemosPlanCoreBundle/components/DpVideoPlayer'
+import DpVideoPlayer from '@DpJs/components/core/DpVideoPlayer'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
 import layers from '@DemosPlanMapBundle/store/Layers'
 import Map from '@DemosPlanMapBundle/components/publicdetail/Map'
