+++ demosplan/DemosPlanMiscContentBundle/Resources/client/js/bundles/signLanguage.js
@@ -11,7 +11,7 @@
  * This is the entrypoint for sign_language.html.twig
  */
 
-import DpVideoPlayer from '@DemosPlanCoreBundle/components/DpVideoPlayer'
+import DpVideoPlayer from '@DpJs/components/core/DpVideoPlayer'
 import { initialize } from '@DemosPlanCoreBundle/InitVue'
 
 const components = { DpVideoPlayer }
